+++ src/main/kotlin/spartacodingclub/nbcamp/kotlinspring/project/team4ighting/spring4gamer/infra/s3/fileupload/controller/FileUploadController.kt
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
-import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.infra.s3.fileupload.dto.S3GetResponseDto
+import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.infra.s3.fileupload.dto.response.S3GetResponseDto
 import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.infra.s3.fileupload.service.FileUploadService
 import spartacodingclub.nbcamp.kotlinspring.project.team4ighting.spring4gamer.infra.security.MemberPrincipal
 
@@ -26,27 +26,30 @@ class FileUploadController(
             .status(HttpStatus.CREATED)
             .body(fileUploadService.preSignedUrl(file))
 
+
     // 이미지 단건 조회
     @GetMapping("/images")
     fun getFile(
         @AuthenticationPrincipal member: MemberPrincipal,
         @RequestParam file: String,
     ) : ResponseEntity<String> =
+
         ResponseEntity
             .status(HttpStatus.OK)
             .body(fileUploadService.getFile(file))
+
 
     // 이미지 전체 조회
     @GetMapping("/{bucket}/**")
     fun find(
         @PathVariable bucket: String,
         request: HttpServletRequest
-    ): ResponseEntity<S3GetResponseDto> {
-        val split = request.requestURI.split("/s3/$bucket")
-        val prefix = if (split.size < 2) "" else split[1].substring(1)
-        val s3GetResponseDto = fileUploadService.find(bucket, prefix)
-        return ResponseEntity.ok(s3GetResponseDto)
-    }
+    ): ResponseEntity<S3GetResponseDto> =
+
+        ResponseEntity
+            .status(HttpStatus.OK)
+            .body(fileUploadService.find(bucket, request))
+
 
     // 이미지 삭제
     @DeleteMapping("/images")
@@ -54,6 +57,6 @@ class FileUploadController(
         @AuthenticationPrincipal member: MemberPrincipal,
         @RequestParam file: String,
     ): ResponseEntity<Unit> =
-        ResponseEntity.status(HttpStatus.OK).body(fileUploadService.delete(file))
 
+        ResponseEntity.status(HttpStatus.OK).body(fileUploadService.delete(file))
 }
